requirements/base.txt
arrow==0.12.1
blessed==1.14.2
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-elasticsearch-dsl==0.4.4
django-elasticsearch-dsl-drf==0.8.3
django-excel==0.0.10
django-fcm==0.1.1
django-nine==0.1.13
django-picklefield==1.0.0
django-q==0.9.4
django-q-email==1.0.0
django-rest-auth==0.9.3
django-solo==1.1.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
et-xmlfile==1.0.1
fcm-django==0.2.16
idna==2.6
ipaddress==1.0.19
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.11.1
lml==0.0.1
oauthlib==2.0.6
openpyxl==2.5.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyexcel==0.5.7
pyexcel-io==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.6
pyexcel-xlsx==0.5.6
pyfcm==1.4.5
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-aws4auth==0.9
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
texttable==1.2.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
